+++ providers/azure/azure_data_lake_gen2.go
@@ -16,7 +16,6 @@ import (
 	"github.com/go-kit/log/level"
 	"github.com/pkg/errors"
 	"github.com/thanos-io/objstore"
-	"gopkg.in/yaml.v2"
 )
 
 type DataLakeGen2Bucket struct {
@@ -313,16 +312,13 @@ func (b *DataLakeGen2Bucket) Close() error {
 func NewTestDataLakeGen2Bucket(t testing.TB, component string) (objstore.Bucket, func(), error) {
 	t.Log("Using test Azure data lake gen 2 bucket.")
 
-	conf := &DefaultConfig
+	conf := DefaultConfig
 	conf.StorageAccountName = os.Getenv("AZURE_STORAGE_ACCOUNT")
 	conf.StorageAccountKey = os.Getenv("AZURE_STORAGE_ACCESS_KEY")
 	conf.ContainerName = objstore.CreateTemporaryTestBucketName(t)
+	conf.StorageAccountType = AzStorageAccountType_DataLake
 
-	bc, err := yaml.Marshal(conf)
-	if err != nil {
-		return nil, nil, err
-	}
-	bkt, err := NewBucket(log.NewNopLogger(), bc, component, nil)
+	bkt, err := NewDataLakeGen2Bucket(log.NewNopLogger(), conf, component, nil)
 	if err != nil {
 		t.Errorf("Cannot create Azure storage container:")
 		return nil, nil, err
@@ -330,8 +326,7 @@ func NewTestDataLakeGen2Bucket(t testing.TB, component string) (objstore.Bucket,
 	ctx := context.Background()
 	return bkt, func() {
 		objstore.EmptyBucket(t, ctx, bkt)
-		_, err := bkt.(*DataLakeGen2Bucket).filesystemClient.Delete(ctx, nil)
-		if err != nil {
+		if _, err := bkt.filesystemClient.Delete(ctx, nil); err != nil {
 			t.Logf("deleting bucket failed: %s", err)
 		}
 	}, nil